+++ ui/src/admin/components/QueriesTable.js
@@ -40,13 +40,13 @@ const QueriesTable = ({queries, queriesSort, changeSort, onKillQuery}) => {
           </th>
           <th>Query</th>
           <th
-            style={{width: `${QUERIES_TABLE.colRunning}px`}}
+            style={{width: `${QUERIES_TABLE.colDuration}px`}}
             className={`col--sortable ${timeSortClass}`}
             onClick={() => changeSort(newTimeSort)}
           >
-            Running
+            Duration
           </th>
-          <th style={{width: `${QUERIES_TABLE.colKillQuery}px`}} />
+          <th style={{width: `${QUERIES_TABLE.colKillQuery}px`}}>Status</th>
         </tr>
       </thead>
       <tbody>