+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. [#5831](https://github.com/influxdata/chronograf/pull/5831): Add download button on query management page.
 1. [#5836](https://github.com/influxdata/chronograf/pull/5836): Allow to rename TICKscript.
+1. [#5837](https://github.com/influxdata/chronograf/pull/5837): Add status column to Queries page.
 
 ### Bug Fixes